mise a la norme, clean

+++ srcs_pushswap/ft_toolstack.c
@@ -6,23 +6,21 @@
 /*   By: cammapou <cammapou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/17 12:02:28 by cammapou          #+#    #+#             */
-/*   Updated: 2018/06/09 15:07:37 by cammapou         ###   ########.fr       */
+/*   Updated: 2018/06/14 15:48:33 by cammapou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/push_swap.h"
 
-
-int	ft_min_value(t_list *lst)
+int		ft_min_value(t_list *lst)
 {
 	t_list	*tmp;
-	int			*min;
+	int		*min;
 
 	if (!lst->content)
 		return (0);
 	min = lst->content;
 	tmp = lst->next;
-
 	while (tmp)
 	{
 		if (*min > *(int*)tmp->content)
@@ -35,13 +33,12 @@ int	ft_min_value(t_list *lst)
 int		ft_max_value(t_list *lst)
 {
 	t_list	*tmp;
-	int			*max;
+	int		*max;
 
 	if (!lst->content)
 		return (0);
 	max = lst->content;
 	tmp = lst->next;
-
 	while (tmp)
 	{
 		if (*max < *(int*)tmp->content)
@@ -51,12 +48,12 @@ int		ft_max_value(t_list *lst)
 	return (*max);
 }
 
-size_t ft_stack_len_b(t_list *lst_b)
+int		ft_stack_len_b(t_list *lst_b)
 {
-	size_t len;
+	size_t	len;
 
 	len = 0;
-	while(lst_b)
+	while (lst_b)
 	{
 		len++;
 		lst_b = lst_b->next;
@@ -66,12 +63,12 @@ size_t ft_stack_len_b(t_list *lst_b)
 		return (0);
 }
 
-size_t ft_stack_len_a(t_list *lst_a)
+int		ft_stack_len_a(t_list *lst_a)
 {
-	size_t len;
+	int		len;
 
 	len = 0;
-	while(lst_a)
+	while (lst_a)
 	{
 		len++;
 		lst_a = lst_a->next;
@@ -81,11 +78,10 @@ size_t ft_stack_len_a(t_list *lst_a)
 		return (0);
 }
 
-
-int	ft_stack_de(t_list *lst_a)
+int		ft_stack_de(t_list *lst_a)
 {
 	int			*nbr_tmp;
-	t_list	*tmpstack;
+	t_list		*tmpstack;
 
 	tmpstack = lst_a;
 	nbr_tmp = (int*)tmpstack->content;
